move OnEvent to the event handler
